+++ src/ScottPlot4/ScottPlot/Plottable/BarSeries.cs
@@ -42,7 +42,7 @@ public class BarSeries : IPlottable
         return limits;
     }
 
-    public LegendItem[] GetLegendItems() => Array.Empty<LegendItem>();
+    public LegendItem[] GetLegendItems() => LegendItem.None;
 
     public void ValidateData(bool deep = false) { }