+++ src/Shared/ScottPlotTests/Usings.cs
@@ -1,2 +1,3 @@
 global using NUnit.Framework;
 global using FluentAssertions;
+global using System.Linq;