+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## ScottPlot 4.1.62 (in development)
 * Ellipse and Circle: New plot types demonstrated in the cookbook. (#2413, #2437) _Thanks @bukkideme_
 * Heatmap: Added `FlipVertically` to invert vertical axis of heatmap data (#2444, #2450) _Thanks @Neopentane1_
+* Histogram: Improved support for datasets with low variance (#2464, #2463) _Thanks @Xerxes004_
 
 # Releases